Separated events and event stream
The util directory seems like a good place for private code.
